Mismatch with ACPI definition. (Refactoring ACPI never looked so good...)
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@38283 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/system/kernel/arch/x86/irq_routing_table.h
@@ -45,7 +45,7 @@ struct irq_descriptor {
 typedef struct acpi_prt {
 	uint32			length;
 	uint32			pin;
-	int				address;		// here for 64-bit alignment
+	uint64			address;		// here for 64-bit alignment
 	uint32			sourceIndex;
 	char			source[4];		// pad to 64 bits so sizeof() works in
 									// all cases
